Optimize Grpc instrumentation laoding
