+++ test/IntegrationTests/ModuleTests.Default.NetCore.verified.txt
@@ -5,7 +5,6 @@
   OpenTelemetry.AutoInstrumentation.Loader,
   OpenTelemetry.AutoInstrumentation.StartupHook,
   OpenTelemetry.Exporter.OpenTelemetryProtocol,
-  OpenTelemetry.Instrumentation.GrpcNetClient,
   OpenTelemetry.Instrumentation.Http,
   OpenTelemetry.Instrumentation.Process,
   OpenTelemetry.Instrumentation.Runtime,
